improve overall codebase

+++ src/main/scala/taranis/core/package.scala
@@ -2,6 +2,6 @@ package taranis
 
 package object core {
 
-  type Time = Long
+  type Time = Double
 
 }
